Move role schemas to own file
