+++ src/types/schema/mod.rs
@@ -3,6 +3,7 @@ mod auth;
 mod channel;
 mod guild;
 mod message;
+mod role;
 mod user;
 
 pub use apierror::*;
@@ -10,6 +11,7 @@ pub use auth::*;
 pub use channel::*;
 pub use guild::*;
 pub use message::*;
+pub use role::*;
 pub use user::*;
 
 #[cfg(test)]
